fix tests

+++ redis-client_test.go
@@ -129,13 +129,13 @@ func TestHSetCommand(t *testing.T) {
 
 	res, err := rdb.HSet("myhash", "field1", "foo")
 
-	if res.(int) != 1 {
+	if res.(int64) != 1 {
 		t.Fatalf("Incorrect response: %v", err)
 	}
 
 	res, err = rdb.HDel("myhash", "field1")
 
-	if res.(int) != 1 {
+	if res.(int64) != 1 {
 		t.Fatalf("Incorrect response from hdel command: %v", err)
 	}
